+++ src/game/scenes/game/Game.js
@@ -1,15 +1,19 @@
-import { EventBus } from '../EventBus';
-import { Scene } from 'phaser';
+import Phaser from 'phaser';
+import { EVENTS, EventBus } from '../../EventBus';
 
-export class Game extends Scene {
+
+export class Game extends Phaser.Scene {
+  buttons = {};
 
   constructor() {
     super('Game');
   }
 
   create() {
+    // add background color
     this.cameras.main.setBackgroundColor(0x00ff00);
 
+    // add background image
     this.add.image(512, 384, 'background').setAlpha(0.5);
 
     this.add.text(512, 384, 'INSERT ACTUAL GAME HERE', {
@@ -18,10 +22,6 @@ export class Game extends Scene {
       align: 'center'
     }).setOrigin(0.5).setDepth(100);
 
-    EventBus.emit('current-scene-ready', this);
-  }
-
-  changeScene() {
-    this.scene.start('MainMenu');
+    EventBus.emit(EVENTS.CURRENT_SCENE_READY, this);
   }
 }
